io_handlers errors

+++ software/dimos/src/ui/io_handlers/can_handlers.h
@@ -9,7 +9,7 @@ namespace can_handlers
 constexpr uint16_t TASK_INTERVAL_100HZ = 1000;
 constexpr uint16_t TASK_INTERVAL_1HZ   = 1000;
 #elif USING_dimos
-constexpr uint8_t TASK_INTERVAL_100HZ = 10;
-constexpr uint8_t TASK_INTERVAL_1HZ   = 1000;
+constexpr uint16_t TASK_INTERVAL_100HZ = 10;
+constexpr uint16_t TASK_INTERVAL_1HZ   = 1000;
 #endif
 } // namespace can_handlers
